% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Edit.R
\name{blicc_combine_bins}
\alias{blicc_combine_bins}
\title{Combines bins across a range into a single length frequency bin}
\usage{
blicc_combine_bins(blicc_ld, interval_range)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{interval_range}{The range of bins identified by their LLB that are to
be combined (lower bin and upper bin).}
}
\value{
The same data object blicc_ld but with the bin adjustment.
}
\description{
The frequencies are summed over an interval range which is converted
into a single bin. This may be useful for smoothing peak outliers across bins.
The resulting frequency bins will have variable width.
}
