% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{blicc_expected_catches}
\alias{blicc_expected_catches}
\title{Generate expected proportional catch numbers for each gear.}
\usage{
blicc_expected_catches(blicc_rp)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}
}
\value{
A tibble of expected catches and residuals for each gear and MCMC
iteration
}
\description{
The function returns a tibble of the expected relative catch number for each
gear for the provided parameters. The gear values add up to 1, so for single
gear fisheries, this is will return 1; therefore this is only useful in
multi-gear fisheries.
}
\details{
As well as the expected catch, the fit residuals are returned as a
diagnostic. For the mpd fit, the function returns a tibble with the number of
rows equal to the number of gears. For the MCMC fit, it returns the number of
MCMC iterations multiplied by the number of gears.
}
\examples{
blicc_expected_catches(trgl_rp) |>
  dplyr::group_by(gear) |>
  dplyr::summarise(resid=mean(resid))

}
