% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blicc_fit}
\alias{blicc_fit}
\title{Fit a Bayesian length interval catch curve to a length frequency data sample}
\usage{
blicc_fit(ld, ptargn = 2000, pwup = 1000, pthin = 1, pchain = 4)
}
\arguments{
\item{ld}{A standard data list (see function BLICC_dat)}

\item{ptargn}{target draw for the MCMC}

\item{pwup}{warm up iterations for the Stan MCMC}

\item{pthin}{thinning for the MCMC draws - only set above 1 if there is significant autocorrelation}

\item{pchain}{number of chains to run in parallel}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
The model assumes constant recruitment, double-sided normal selectivity, a single natural
mortality parameter, and Gamma-distributed von Bertalanffy growth to define expected
numbers of fish in pre-defined length bins. The model assumes a negative binomial likelihood
function to fit to a single length frequency sample. Selectivity, mortality, asymptotic mean
length and scale parameters are fitted. In addition, the spawning potential ratio is generated
which ca be used to indicate the stock status.
}
