% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{blicc_get_efq}
\alias{blicc_get_efq}
\title{The expected length frequency}
\usage{
blicc_get_efq(Linf, Galpha, Mk, Fk, Smx, Ss1, Ss2, blicc_ld)
}
\arguments{
\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Fk}{Fishing mortality (time in units of the growth rate K) to
be multiplied by selectivity to get fishing mortality in each length bin}

\item{Smx}{Mode of the normal selectivity function (full selectivity)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{blicc_ld}{A standard data list created by \code{blicc_dat}}
}
\value{
A vector of the expected length bin frequency
}
\description{
For a set of parameter values, returns the expected catch in each length bin
based on the BLICC model. This is the same as the blicc_get_expected()
function, but only returns the expected length frequency.
}
\examples{
lfq <- blicc_get_efq(Linf = 32, Galpha = 100, Mk = 1.5, Fk = 1.5, Smx = 24,
                     Ss1 = 0.1, Ss2 = 0.001, blicc_ld = eg_ld)
plot(y = lfq, x = eg_ld$Len)

}
