% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{blicc_get_expected}
\alias{blicc_get_expected}
\title{Expected length interval selectivity, mortality, survival and frequency}
\usage{
blicc_get_expected(Linf, Galpha, Mk, Fk, Smx, Ss1, Ss2, blicc_ld)
}
\arguments{
\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Fk}{Fishing mortality (time in units of the growth rate K) to
be multiplied by selectivity to get fishing mortality in each length bin}

\item{Smx}{Mode of the normal selectivity function (full selectivity)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{blicc_ld}{A standard data list created by \code{blicc_dat}}
}
\value{
A tibble of expected values for each length bin
}
\description{
For a set of parameter values, returns the expected selectivity, mortality,
survival, and catch in each length bin based on the BLICC model as a tibble.
}
\details{
The model and provided parameter values are used to calculate the expected
values for selectivity, survival, relative population numbers,
total mortality and expected length frequency for each length bin.
Input parameters are point estimates (reals). The standard data list
provides the model dimensions and is produced by the function
blicc_dat().
}
\examples{
blicc_get_expected(Linf = 32, Galpha = 100, Mk = 1.5, Fk = 1.5, Smx = 24,
                   Ss1 = 0.1, Ss2 = 0.001, blicc_ld = eg_ld)

}
