% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Tables.R
\name{blicc_impact}
\alias{blicc_impact}
\title{Impact from the removal of each gear on the YPR and SPR in the specified
period.}
\usage{
blicc_impact(blicc_rp)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}
}
\value{
A tibble with the changes in YPR and SPR from the removal of each
gear for the scenario in blicc result list.
}
\description{
The change in YPR and SPR are calculated for each parameter set to produce a
tibble of the YPR and SPR change from the removal of each gear in sequence.
This is only useful if there is more than one gear.
}
\examples{
blicc_impact(blicc_ref_pts(blicc_mpd(trgl_ld), trgl_ld))
  
}
