% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blicc_ini}
\alias{blicc_ini}
\title{Default start parameters centred on priors}
\usage{
blicc_ini()
}
\value{
A list of initial parameter values for the BLICC Stan model
}
\description{
Default start parameters centred on priors
}
