% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_stan.R
\name{blicc_mcmc_ini}
\alias{blicc_mcmc_ini}
\title{Provides start parameters to the number of chains required for the MCMC run}
\usage{
blicc_mcmc_ini(pchain = 4, par = NULL)
}
\arguments{
\item{pchain}{A number indicating the number of chains. Default is 4}

\item{par}{A list of start parameter values (optional)}
}
\value{
A list of start parameters equal to the number of chains for
the BLICC Stan model
}
\description{
Provides a list of start values for the BLICC model MCMC using function
\code{blicc_ini} if no parameters are provided.
}
