% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blicc_period_filter}
\alias{blicc_period_filter}
\title{Returns data list from \link{blicc_dat} with only the selected time period
included}
\usage{
blicc_period_filter(blicc_ld, time_period)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{time_period}{The names or integer vector of the time periods being
selected}
}
\value{
The data object blicc_ld subset for the new time period.
}
\description{
Subsets the data list with only selectivity components, gears, frequencies
and parameters relevant to the defined time period.
}
