% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Edit.R
\name{blicc_population_filter}
\alias{blicc_population_filter}
\title{Returns data list from \link{blicc_dat} with only the selected populations
included}
\usage{
blicc_population_filter(blicc_ld, population)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{population}{The names or an integer vector of the populations (time
period/growth groups) being selected.}
}
\value{
The data object blicc_ld subset for the new populations
(growth groups / time period).
}
\description{
Subsets the data list with only selectivity components, gears, frequencies
and parameters relevant to the defined populations.
}
