% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Tables.R
\name{blicc_prior}
\alias{blicc_prior}
\title{Return a tibble containing a summary of the fishblicc priors being applied}
\usage{
blicc_prior(blicc_ld)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}
}
\value{
A tibble describing the priors being applied.
}
\description{
This function provides a summary of the priors contained in a
fishblicc data object in a tibble. This can be used to inspect the model's
assumptions or to create a table in a report.
}
\examples{
blicc_prior(eg_ld)

}
