% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Stan.R
\name{blicc_selfun}
\alias{blicc_selfun}
\title{Sets data object from \code{\link[=blicc_dat]{blicc_dat()}} to have a new selectivity functions}
\usage{
blicc_selfun(blicc_ld, Gear, sel_fun, model_name = NA)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{Gear}{The gears for which the selectivity functions are being
changed}

\item{sel_fun}{The selectivity functions that gears are being set to either
as integers or strings.}

\item{model_name}{A string for a replacement model name in the data
object. Optional.}
}
\value{
The data object blicc_ld with the new selectivities
}
\description{
Checks the gear and selectivity are valid, then sets up the data object
so that the parameters and parameter references for the selectivity
functions are valid. New references are replaced in the data object,
which is then returned.
}
\examples{
new_ld <- blicc_selfun(eg_ld, Gear=1, sel_fun="logistic", model_name = "Logistic Selectivity")

}
