% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_catchLN}
\alias{blip_catchLN}
\title{Set the relative catch log-normal likelihood standard deviation}
\usage{
blip_catchLN(blicc_ld, sigma)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{sigma}{The standard deviation for the lognormal likelihood}
}
\value{
The data object blicc_ld but with the value for \code{polCs} changed.
}
\description{
The relative catches for multigear fisheries are fitted using a lognormal. The
default is set at around 5\%, but lower values may be necessary to ensure the
catches are fitted particularly when large number of length observations are
used. Has no effect in single gear fisheries.
}
\examples{
new_ld <- blip_catchLN(trgl_ld, sigma = 0.01)

}
