% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Stan.R
\name{blip_LH_param}
\alias{blip_LH_param}
\title{Sets data object from \code{\link[=blicc_dat]{blicc_dat()}} to have a new selectivity functions}
\usage{
blip_LH_param(
  blicc_ld,
  a = NA,
  b = NA,
  L50 = NA,
  L95 = NA,
  ma_L = NA,
  wt_L = NA,
  model_name = NA,
  set_defaults = FALSE
)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{a}{The length-weight parameter: a*L^b. Not used in model fitting, but
for plots etc. Optional.}

\item{b}{The length-weight exponent (a*L^b, usually close to 3.0) Optional.}

\item{L50}{The length at 50\% maturity, often referred to as "first" maturity,
and primarily applies to females. Must be less than Linf, usually around
0.66 Linf, which is assumed if it is not provided. Optional.}

\item{L95}{The length at 95\% maturity. Must be greater than L50 and less
than Linf. A small increment is added L50 if it is not provided. Optional.}

\item{ma_L}{Mature biomass for each length bin. Optional.}

\item{wt_L}{Weight (biomass) for each length bin. Optional.}

\item{model_name}{A string for a replacement model name in the data
object. Optional.}

\item{set_defaults}{Logical indicating whether to set defaults or not if
parameters are NA}
}
\value{
The data object blicc_ld with the new life-history parameters
}
\description{
Checks the gear and selectivity are valid, then sets up the data object so
that the parameters and parameter references for the selectivity functions are
valid. New references are replaced in the data object, which is then returned.
}
\examples{
new_ld <- blip_LH_param(eg_ld, a=1.2e-5, b=2.95, model_name="Alternative LW")

}
