% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_NBphi}
\alias{blip_NBphi}
\title{Set the \code{NB_phi} prior hyper-parameters in the data object}
\usage{
blip_NBphi(blicc_ld, lNBphi)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{lNBphi}{A numeric vector of double containing the mu and sigma for the
prior log-normal.}
}
\value{
The data object blicc_ld but with the prior for \code{NBphi} changed.
}
\description{
The \code{NB_phi} log-normal prior is updated with a mean (mu) and standard
deviation (sigma). NB_phi controls the negative binomial over-dispersion
compared to the Poisson distribution. It is sometimes useful to reduce the
NB_phi lognormal standard deviation of the hyper-parameter to below the 0.5
default to avoid variation in the data being interpreted as observation error when it
can be explained better by the model.
}
\details{
The variance for the negative binomial is \code{ Var = mu + mu^2 / NB_phi}
compared to \code{Var = mu} for the Poisson distribution.
}
\examples{
new_ld <- blip_NBphi(gillnet_ld, lNBphi = c(log(100), 0.05))

}
