% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_normalize_mix}
\alias{blip_normalize_mix}
\title{Normalize selectivity mixture weights, so the maximum mixture weight is 1.0}
\usage{
blip_normalize_mix(blicc_ld)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}
}
\value{
The data list \code{blicc_ld} with normalized weights
}
\description{
The function checks each gear that has mixtures that the maximum weight is
1.0. If it is higher than 1.0, the selectivity with the highest weight is
moved to the base component using \link{blip_main_sel}.
}
