% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_sel}
\alias{blip_sel}
\title{Set a selectivity function's prior hyper-parameters to the given values}
\usage{
blip_sel(blicc_ld, sel_indx, loc, lslope = NULL)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{sel_indx}{Single integer indexing a selectivity function}

\item{loc}{Single location parameter for the selectivity function}

\item{lslope}{Vector of prior selectivity log slope parameters manually set
for the specified selectivity function. Optional.}
}
\value{
The data object \code{blicc_ld} but with the new selectivity priors set
}
\description{
Each parameter set is for a single selectivity function indexed by \code{sel_indx}.
The location parameter (\code{loc}) is either the 50\% selectivity for the logistic
or the mode of the particular normal selectivity function, and is required.
}
\details{
The \code{lslope} parameter is a single or double log value of the slope for the
relevant function. This is either the log of the logistic slope parameter, or
the log of the reciprocal of the variance for the normal functions. For the
double-sided normal, two parameters must be provided and for all other
functions only one parameter. If \code{lslope} is not provided, a default slope of
-4.5 is applied if the slope parameters have not already been set. If values
are already present these are conserved. Priors for mixture weights, if a
mixture is used, must be set in \link{blip_mix}.
}
\examples{
new_ld <- blip_sel(gillnet_ld, sel_indx = 1, loc = 30, lslope = -2.8)

}
