% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_set_sel}
\alias{blip_set_sel}
\title{Sets a selectivity function's prior hyper-parameters to particular values}
\usage{
blip_set_sel(blicc_ld, sel_indx, loc, lslope = NULL)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{sel_indx}{Single integer indexing a selectivity function}

\item{loc}{Single location parameter for the selectivity function}

\item{lslope}{Vector of prior selectivity log slope parameters manually set for
the model. Only one gear at a time can be specified to use this. Parameters
must be in the correct order: in function order, with each function having
a location parameter followed by one or two steepness parameters.}
}
\value{
The data object \code{blicc_ld} but with the new selectivity priors set
}
\description{
Each parameter set is for a single selectivity function indexed by
\code{sel_indx}. Although the hyper-parameters can be set directly in the data
object, this function does the indexing making it a little easier. The
location parameter is set as a length measure. The slope parameters are set
as log values.
}
\details{
The location parameter (\code{loc}) is either the 50\% selectivity for the logistic
or the mode of the particular normal selectivity function. The \code{lslope}
parameter is a single log value of the slope for the relevant function. This
is either the log of the logistic slope parameter, or the log of the
reciprocal of the variance for the normal functions. For the double-sided
normal, two parameters must be provided and for all other functions only one
parameter. If \code{lslope} is not provided, a default slope of -4.5 is applied if
the slope parameters have not already set. If values are already present
these are conserved.
}
