% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_slim}
\alias{eg_slim}
\title{Example stanfit object from the BLICC model (Stan MCMC)}
\format{
A \code{stanfit} object: variables can be obtained using \code{rstan::extract} and
other methods. see packages \code{rstan} and \code{posterior} for details.
}
\source{
Data are simulated.
}
\usage{
eg_slim
}
\description{
A stanfit object from fitting the BLICC model using \link{blicc_fit} to the "eg_ld"
data set. The draws can be extracted and reference points
calculated using function \link{blicc_ref_pts}. The MCMC simulations are also
accessible through a variety of \code{rstan} and \code{posterior} functions.
eg_slim <- blicc_fit(eg_ld, nwarmup=500, thin=2)
}
\keyword{datasets}
