% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_PR.R
\name{F01_solve}
\alias{F01_solve}
\title{Find the standard yield-per-recruit fishing mortality reference point F0.1}
\usage{
F01_solve(Smx, Linf, Galpha, Mk, Ss1, Ss2, blicc_ld, glq)
}
\arguments{
\item{Smx}{Mode of the normal selectivity function (full selectivity)}

\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{blicc_ld}{A standard data list created by \code{blicc_dat}}

\item{glq}{a list of \code{nodes} and \code{weights} for the
Gauss-Laguerre rule (see \code{statmod::gauss.quad()})}
}
\value{
F0.1 per-recruit fishing mortality reference point
}
\description{
Estimates an F0.1 reference point consistent with parameters and YPR slope
10\% of initial slope. The function solves for F0.1 using a simple
linear approximation to slope at the origin and at each point on the curve.
}
