% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc-package.R
\docType{package}
\name{fishblicc-package}
\alias{fishblicc-package}
\alias{fishblicc}
\title{The 'fishblicc' package.}
\description{
Procedures to fit catch curves to length frequency data based
using length intervals.
}
\details{
The package supports fitting a Bayesian length interval catch curve,
with selectivity modelled as an optional mixture of logistic and normal
functions. The package is designed to be used with multiple gears with
different selectivities fishing the same stock.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulAHMedley/fishblicc}
  \item Report bugs at \url{https://github.com/PaulAHMedley/fishblicc/issues}
}

}
\author{
\strong{Maintainer}: Paul A.H. Medley \email{paulahmedley@gmail.com}

Authors:
\itemize{
  \item Paul A.H. Medley \email{paulahmedley@gmail.com}
}

}
\keyword{internal}
