% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_PR.R
\name{fSPR}
\alias{fSPR}
\title{Spawning biomass per recruit with variable fishing mortality}
\usage{
fSPR(Fk, Galpha, Gbeta, Mk, RSel, Len, mb, glq)
}
\arguments{
\item{Fk}{Fishing mortality (per unit growth rate K)}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Gbeta}{Rate parameter for the Gamma distribution growth variability
(=Galpha/Linf)}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{RSel}{Vector of selectivity from function \code{Rsel_dsnormal} or
another source}

\item{Len}{Vector of lower length boundaries for the length bins}

\item{mb}{Vector of mature biomass from \code{mature_biomass_at_length} function}

\item{glq}{a list of \code{nodes} and \code{weights} for the
Gauss-Laguerre rule (see \code{statmod::gauss.quad})}
}
\value{
spawning biomass per recruit
}
\description{
Used to evaluate the SPR for different values of Fk.
}
