% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_PR.R
\name{FSPR_solve}
\alias{FSPR_solve}
\title{Solve for a fishing mortality which produces the target SPR}
\usage{
FSPR_solve(Smx, Linf, Galpha, Mk, Ss1, Ss2, tarSPR, blicc_ld, glq)
}
\arguments{
\item{Smx}{Mode of the normal selectivity function (full selectivity)}

\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{tarSPR}{target SPR reference point: usually 0.2, 0.3 or 0.4}

\item{blicc_ld}{A standard data list created by \code{blicc_dat}}

\item{glq}{a list of \code{nodes} and \code{weights} for the
Gauss-Laguerre rule (see \code{statmod::gauss.quad})}
}
\value{
Spawning potential ratio per-recruit fishing mortality consistent
with parameters and reference point
}
\description{
The function finds the fishing mortality that gives the \code{tarSPR} spawning
potential ratio (SPR).
}
\details{
calculates spawning biomass per recruit (Spawning Potential Ratio / SPR).
This should generally work for sensible reference point target.
However, note that fishing mortality estimate
may not be finite, dependent on the selectivity.
}
