% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_PR.R
\name{fYPR2}
\alias{fYPR2}
\title{Yield per recruit with variable selectivity}
\usage{
fYPR2(Smx, Fk, Galpha, Gbeta, Mk, Ss1, Ss2, Len, LMP, weight, glq)
}
\arguments{
\item{Smx}{Mode of the normal selectivity function (full selectivity)}

\item{Fk}{Fishing mortality (per unit growth rate K)}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Gbeta}{Rate parameter for the Gamma distribution growth variability
(=Galpha/Linf)}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{Len}{Vector of lower length boundaries for the length bins}

\item{LMP}{Vector of length bin mid-points to calculate selectivity}

\item{weight}{Vector of weights from \code{weight_at_length} function}

\item{glq}{a list of \code{nodes} and \code{weights} for the
Gauss-Laguerre rule (see \code{statmod::gauss.quad})}
}
\value{
yield-per-recruit
}
\description{
Used to evaluate the YPR for different values of Fk and Smx.
}
