% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gillnet_ld}
\alias{gillnet_ld}
\title{Example small sample of length frequency data from a gillnet fishery}
\format{
A list of 26 data vectors used in the BLICC model:
\describe{
\item{model_name}{Model and/or fit and/or species name}
\item{NQ}{Number of length frequency data sets}
\item{NG}{Number of gears}
\item{NS}{Number of selectivity functions}
\item{NT}{Number of time periods}
\item{NF}{Number of fishing mortality estimates}
\item{NB}{Number of bins in the length frequency}
\item{NP}{Number of selectivity function parameters}
\item{NM}{Number of mixture selectvity models}
\item{fqname}{Length frequency data set names}
\item{gname}{Fishing gear names}
\item{tpname}{Time period names}
\item{LLB}{Vector of length bin lower boundaries}
\item{LMP}{Length bin mid-points, used for selectivity and plotting}
\item{fq}{List of length frequency vectors, one for each gear}
\item{Gi}{Integer vector linking length frequency data sets to gears}
\item{Ti}{Integer vector linking length frequency data sets to time periods}
\item{prop_catch}{Vector of proportional catch by gear of length NF}
\item{Fkq}{Integer vector of length frequency data sets index linking to fishing mortality}
\item{fSel}{Index of selectivity function to use for each selectvity model}
\item{GSbase}{Index of based selectivity model for each gear}
\item{GSmix1}{Index of start and end mixtures in GSmix2: 0 implies no mixture}
\item{GSmix2}{Index of all selectivity mixture models referenced by gear in GSmix1}
\item{poLinfm}{Linf: Mean prior normal parameter}
\item{poLinfs}{Linf: sd prior normal parameter}
\item{polGam}{Galpha: Mean prior normal log parameter}
\item{polGas}{Galpha: sd prior normal log parameter}
\item{a}{Length-weight parameter in a L^b. Optional}
\item{b}{Length-weight parameter in a L^b.}
\item{L50}{Logistic parameter for 50\% maturity}
\item{Ls}{Logistic steepness parameter for the maturity ogive}
\item{wt_L}{Vector of average weight in each length bin}
\item{ma_L}{Vector of mature biomass in each length bin}
\item{polMkm}{Mk: Mean prior normal log parameter}
\item{polMks}{Mk: sd prior normal log parameter}
\item{polFkm}{Fk: Mean prior normal log parameter}
\item{polFks}{Fk: sd prior normal log parameter}
\item{sp_i}{Integer index of the parameter start in polSm for each selectivity function}
\item{sp_e}{Integer index of the parameter end in polSm for each selectivity function}
\item{polSm}{Sm: Mean priors normal log selectivity parameter including mixture weights at the end}
\item{polSs}{Sm: sd priors normal log selectivity parameter including mixture weights at the end}
\item{polNB_phim}{NB_phi: Mean prior log normal parameter}
\item{polNB_phis}{NB_phi: sd prior log normal parameter}
\item{polCs}{Catch: sd prior log normal parameter on relative catch}
\item{NK}{Number of nodes in the Gauss-Laguerre quadrature}
\item{gl_nodes}{Gauss-Laguerre quadrature nodes}
\item{gl_weights}{Gauss-Laguerre quadrature weights}
}
}
\source{
Data are simulated.
}
\usage{
gillnet_ld
}
\description{
A data list created using simulated length frequency data from the
function \link{blicc_dat}:
}
\details{
gillnet_ld <- fishblicc::blicc_dat(model_name = "Simulated Gillnet Data",
LLB = seq(14, 62, by=2),
fq=list(Gillnet = c(0, 2, 7, 18, 32, 56, 115, 155, 227, 240, 246, 249,
210, 158, 106, 73, 43, 28, 7, 6, 4, 3, 0, 1, 0)),
sel_fun = "normal",
gear_names = "Gillnet",
Linf=c(64, 3),
Mk=1.55,
a=1.0e-4, b=2.95, L50=34, L95=43)
}
\keyword{datasets}
