% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{mature_biomass_at_length}
\alias{mature_biomass_at_length}
\title{Mature biomass at length}
\usage{
mature_biomass_at_length(blicc_ld)
}
\arguments{
\item{blicc_ld}{A standard data list created by \code{blicc_dat}}
}
\description{
Mature biomass is calculated based on the length-weight relationship and
the maturity ogive. In terms of calculating SPR and reference points,
only the exponent parameter b and maturity ogive have any effect.
The length-weight parameter a is only used for scaling purposes
and need not be provided.
}
\examples{
ld <- blicc_dat(LLB = 10:35, a=1.0e-4, b=2.95, L50=24,
                fq=c(0,1,2,5,26,70,72,66,36,34,25,24,
                     20,10,12,5,3,5,6,4,2,0,2,0,1,0), Linf=c(32, 3))
mb <- mature_biomass_at_length(ld)
plot(y=mb, x=ld$LMP, type="l")

}
