% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{NinInterval}
\alias{NinInterval}
\title{Calculate the population relative numbers-at-length within length bins}
\usage{
NinInterval(Surv, Zki, pstream__ = <pointer: 0x00007ffea6d9df40>)
}
\arguments{
\item{Zki}{Total mortality within each bin (time in units of the
growth rate K)}
}
\value{
A vector of relative numbers of fish within each length bin.
}
\description{
Place holder for the rstan::expose_stan_functions
}
