% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_efq}
\alias{plot_efq}
\title{Plot expected length frequencies for a range of fishing mortality and
selectivity reference points}
\usage{
plot_efq(blicc_rp, gear = NULL)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{gear}{Specifies the gear(s) to plot as an integer index or full name}
}
\value{
ggplot geom object for plotting expected length frequencies by SPR
level
}
\description{
The facet plot estimates the length frequency data expected for a range of
fishing mortality and selectivity to show how the length frequency data might
respond to changes in mortality and/or selectivity. Fishing mortalities
tested are SPR 20\% and SPR 40\% and selectivity to obtain SPR 40\% and the
maximum yield.
}
\details{
The graphs show the expected values and 80\% credible intervals
compared to the observations. They can be used to assess whether length
frequencies should be able to detect changes in fishing mortality to these
different levels. In many cases, differences between frequencies can be
subtle and the precision of data collection may not be sufficient to detect
such changes reliably.  Dependent on the selectivity, reference points may
not exist. It may not be possible to achieve an SPR level without changing
multiple parameters, whereas these scenarios are restricted to only
changing either the fishing mortality or the selectivity function central
locations. If reference points do not exist, they are not plotted, so some
graphs may be blank except for the data. The graphs show the number of MCMC
draws for which the reference points exist in each case.

If the apparent changes in length frequencies between the plots is small,
relying on length frequency data alone may not be advisable.

If selectivity mixtures are used, all location parameters for the mixtures
will be shifted equally by default. This may not make a lot of sense in
practice and more realistic scenarios may need to be constructed to test
possible management actions.
}
\examples{
plot_efq(eg_rp)

}
