% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_expected_frequency}
\alias{plot_expected_frequency}
\title{Plot the observed and expected frequency data}
\usage{
plot_expected_frequency(blicc_rp, blicc_lx, blicc_ld)
}
\arguments{
\item{blicc_rp}{A posterior draws and reference points tibble
from \code{blicc_ref_pts} function.}

\item{blicc_lx}{A tibble from \code{blicc_expect_len} function.}

\item{blicc_ld}{A data list from the \code{blicc_dat} function.}
}
\value{
ggplot geom object plotting observed and expected frequency
}
\description{
The graph shows the observed frequency and the expected frequency together
with the 80\% credible intervals for the estimate as well as 80\% credible
intervals for the observations. The latter takes into account the negative
binomial observation error, so on average 8 out of ten observations would be
expected within these bounds.
}
\examples{
eg_lx <- blicc_expect_len(eg_rp, eg_ld)
plot_expected_frequency(eg_rp, eg_lx, eg_ld)

}
