% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_FkF40_density}
\alias{plot_FkF40_density}
\title{Plot fishing mortality relative to the SPR 40\% reference point (Fk/F40)}
\usage{
plot_FkF40_density(blicc_rp)
}
\arguments{
\item{blicc_rp}{A posterior draws and reference points tibble
from \code{blicc_ref_pts} function.}
}
\value{
ggplot geom object for plotting
}
\description{
Plot fishing mortality relative to the SPR 40\% reference point (Fk/F40)
}
\examples{
plot_FkF40_density(eg_rp)

}
