% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot the observed and expected length frequency data}
\usage{
plot_posterior(
  blicc_rp,
  gear = "All",
  time_period = "All",
  growth_group = "All"
)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{gear}{Specifies the gear(s) to plot as an integer index or full name}

\item{time_period}{Specifies the period(s) to plot as an integer index or
full name}

\item{growth_group}{Specifies the growth group(s) to plot as an integer index or
full name}
}
\value{
ggplot geom object plotting observed and expected frequency
}
\description{
The graph shows the observed length frequency and the posterior predicted
frequency for the specified gears. The plots include the 80\% credible
intervals for the estimate (if an MCMC is provided) and 80\% credible
intervals for the observations. The latter takes into account the negative
binomial observation error, so on average 8 out of ten observations would be
expected within these bounds.
}
\examples{
plot_posterior(trgl_rp)

}
