% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity and 80\% credible interval, if available, by length.}
\usage{
plot_selectivity(blicc_rp, gear = "All", plotFk = FALSE)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{gear}{Specifies the gear(s) to plot as an integer index or full name}

\item{plotFk}{Whther to plot the graph as fishing mortality at length or
selectivity (default)}
}
\value{
ggplot geom object for plotting
}
\description{
The selectivity at length is plotted for the specified gears. All gears, or
specified gears, are plotted on a single graph. These curves have to be
multiplied by the fishing mortality (Fk) parameters to obtain the relative
fishing mortality at length. The 80\%CI is only available for the MCMC fit.
}
\examples{
plot_selectivity(trgl_rp)

}
