% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_SPR_contour}
\alias{plot_SPR_contour}
\title{Spawning potential ratio surface contour plot}
\usage{
plot_SPR_contour(blicc_rp, blicc_ld)
}
\arguments{
\item{blicc_rp}{A posterior draws and reference points tibble
from \code{blicc_ref_pts} function.}

\item{blicc_ld}{A data list from the \code{blicc_dat} function.}
}
\value{
ggplot geom object for plotting
}
\description{
The plot shows the SPR surface plotted against the fishing mortality (Fk) and
the full selectivity (Smx). The current state of the stock is marked as a
point and the SPR 40\% reference point is marked as a line.
All values are calculated as means and uncertainty is not represented to
maintain simplicity. The reference line is plotted if it exists within a
reasonable range of the current estimated status.
}
\examples{
plot_SPR_contour(eg_rp, eg_ld)

}
