% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_SPR_density}
\alias{plot_SPR_density}
\title{Plot the estimated spawning potential ratio probability density, if
available.}
\usage{
plot_SPR_density(blicc_rp)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}
}
\value{
ggplot geom object for plotting
}
\description{
The SPR estimate as a percentage is converted to a probability density plot
with 20\% and 40\% reference point. This plot requires the MCMC to have been
run. For multiple time periods, densities are overlaid.
}
\examples{
plot_SPR_density(trgl_rp)

}
