% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{RNinInterval}
\alias{RNinInterval}
\title{Calculate the population relative numbers-at-length within length bins}
\usage{
RNinInterval(surv, Zki)
}
\arguments{
\item{surv}{Survival to the lower boundary of each length bin
(see \code{RSurvival_est()})}

\item{Zki}{Total mortality within each bin (time in units of the
growth rate K)}
}
\value{
A vector of relative numbers of fish within each length bin.
}
\description{
The relative numbers in each length bin is based on the definite integral
over the bin growth transition time:
N_i = (S_i - S_{i+1})/Z_{ki}
}
\examples{
glq <- statmod::gauss.quad(90, kind = "laguerre", alpha = 0.0)
S <- RSurvival_Est(glq$nodes, glq$weights, Len=15:55, Zki=c(rep(1.5, 10),
                   rep(3, 31)), 100, 100/50)
P <- RNinInterval(S, Zki=c(rep(1.5, 10), rep(3, 31)))
plot(y=P, x=15:55, type="l")

}
