% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{Rsel_dsnmix}
\alias{Rsel_dsnmix}
\title{Calculate double-sided normal mixture selectivity for a length vector}
\usage{
Rsel_dsnmix(Sp, LMP)
}
\arguments{
\item{Sp}{Vector (length 7) of the selectivity parameters (mode1, mode2,
weight2, then left  and right side 1/sigma^2 for the first and then second)
normal mixture.}

\item{LMP}{A vector of lengths (usually mid-points for length bins)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A vector of selectivity values varying from 0.0 to 1.0
}\if{html}{\out{</div>}}
}
\description{
Two double-sided normal form a mixture allowing two modes in the selectivity.
See \link{Rsel_dsnormal} description. In addition, the model takes a weight
parameter to fit the height of the second curve relative to the first curve.
}
\examples{
Sel <- Rsel_dsnmix(Sp = c(40, 25, 0.5, 0.01, 0.02, 0.1, 0.1), LMP = seq(15.5, 55.5, by=1.0))
plot(y=Sel, x=seq(15.5, 55.5, by=1.0))

}
