% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{Rsel_normal}
\alias{Rsel_normal}
\title{Calculate a normal selectivity curve for a length vector}
\usage{
Rsel_normal(Sp, LMP)
}
\arguments{
\item{Sp}{Vector (length 2) of the selectivity parameters (mode, and
1/sigma^2) for the normal selectivity function.}

\item{LMP}{A vector of lengths (usually mid-points for length bins)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A vector of selectivity values varying from 0.0 to 1.0
}\if{html}{\out{</div>}}
}
\description{
A normal takes one slope (1/sigma^2) parameter for the both sides of the mode
parameter (mu). The resulting selectivity varies from 0 to 1.0, with 1.0
being the mode. It is calculated for each position defined in the length
vector.
}
\examples{
Sel <- Rsel_normal(Sp = c(35, 0.2), LMP = seq(15.5, 55.5, by=1.0))
plot(y=Sel, x=seq(15.5, 55.5, by=1.0))

}
