% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{Rsel_studentt}
\alias{Rsel_studentt}
\title{Calculate a Student's t-distribution selectivity curve for a length vector}
\usage{
Rsel_studentt(Sp, LMP)
}
\arguments{
\item{Sp}{Vector (length 2) of the selectivity parameters (mode and
'degrees of freedom') for the Student's t-distribution selectivity function}

\item{LMP}{A vector of lengths (usually mid-points for length bins)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A vector of selectivity values varying from 0.0 to 1.0
}\if{html}{\out{</div>}}
}
\description{
Student's t-distribution is similar to the normal (\link{Rsel_normal}) but with
heavier tails. A Student's t takes an additional kurtosis parameter
(degrees of freedom: v) as well as a location and scale parameter. As v
decreases, the degree of kurtosis increases. Student's t with high degrees of
freedom resembles the normal. The resulting selectivity varies from 0 to 1.0,
with 1.0 being the mode. It is calculated for each position defined in the
length vector.
}
\examples{
Sel <- Rsel_studentt(Sp = c(35, 0.2, 1), LMP = seq(15.5, 55.5, by=1.0))
plot(y=Sel, x=seq(15.5, 55.5, by=1.0))

}
