% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{RSurvival_Est}
\alias{RSurvival_Est}
\title{Calculate the survival of a fish cohort to sequential length boundaries}
\usage{
RSurvival_Est(node, wt, Len, Zki, Galpha, Gbeta)
}
\arguments{
\item{node}{Nodes for the Gauss-Laguerre quadrature rule}

\item{wt}{Weights for the Gauss-Laguerre quadrature rule}

\item{Len}{Vector of lower-bounds for length intervals}

\item{Zki}{Vector of total mortality (in time units of growth rate
parameter K) to be applied in each length interval.
This vector should be the same length as Len.}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Gbeta}{Beta parameter for the Gamma probability density function
that governs growth variability.}
}
\value{
A vector of proportion surviving to each length bin lower bound.
}
\description{
The model integrates over growth rate variability and sums mortality
piece-wise over intervening length intervals to derive the
survival to each length bin boundary taking into account variation in growth
and varying mortality-at-length. This is analogous to an age based catch
curve model, but for length. Survival can be used to estimate population
numbers-at-length and catch length frequency, among other things.
The model uses the Gauss-Laguerre quadrature rule for integration,
so nodes and weights for this must be provided.
}
\examples{
glq <- statmod::gauss.quad(90, kind = "laguerre", alpha = 0.0)
S <- RSurvival_Est(glq$nodes, glq$weights, Len=15:55,
                   Zki=c(rep(1.5, 10), rep(3, 31)), 100, 100/50)
plot(y=S, x=15:55, type="l")
}
