% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seatrout_ld}
\alias{seatrout_ld}
\title{Example length frequency data set for Guyana seatrout Cynoscion virescens}
\format{
A list of 26 data vectors used in the BLICC model:
\describe{
\item{model_name}{Model and/or fit and/or species name}
\item{NG}{Number of gears}
\item{NS}{Number of selectivity functions}
\item{NF}{Number of fishing mortality estimates}
\item{NB}{Number of bins in the length frequency}
\item{NP}{Number of selectivity function parameters}
\item{NM}{Number of mixture selectvity models}
\item{gname}{Fishing gear names}
\item{LLB}{Vector of length bin lower boundaries}
\item{LMP}{Length bin mid-points, used for selectivity and plotting}
\item{fq}{List of length frequency vectors, one for each gear}
\item{prop_catch}{Vector of proportional catch by gear of length NF}
\item{Fkg}{Integer vector of index for gears link to fishing mortality}
\item{fSel}{Index of selectivity function to use for each selectvity model}
\item{GSbase}{Index of based selectivity model for each gear}
\item{GSmix1}{Index of start and end mixtures in GSmix2: 0 implies no mixture}
\item{GSmix2}{Index of all selectivity mixture models referenced by gear in GSmix1}
\item{poLinfm}{Linf: Mean prior normal parameter}
\item{poLinfs}{Linf: sd prior normal parameter}
\item{polGam}{Galpha: Mean prior normal log parameter}
\item{polGas}{Galpha: sd prior normal log parameter}
\item{a}{Length-weight parameter in a L^b. Optional}
\item{b}{Length-weight parameter in a L^b.}
\item{L50}{Logistic parameter for 50\% maturity}
\item{Ls}{Logistic steepness parameter for the maturity ogive}
\item{wt_L}{Vector of average weight in each length bin}
\item{ma_L}{Vector of mature biomass in each length bin}
\item{polMkm}{Mk: Mean prior normal log parameter}
\item{polMks}{Mk: sd prior normal log parameter}
\item{polFkm}{Fk: Mean prior normal log parameter}
\item{polFks}{Fk: sd prior normal log parameter}
\item{sp_i}{Integer index of the parameter start in polSm for each selectivity function}
\item{sp_e}{Integer index of the parameter end in polSm for each selectivity function}
\item{polSm}{Sm: Mean priors normal log selectivity parameter including mixture weights at the end}
\item{polSs}{Sm: sd priors normal log selectivity parameter including mixture weights at the end}
\item{polNB_phim}{NB_phi: Mean prior log normal parameter}
\item{polNB_phis}{NB_phi: sd prior log normal parameter}
\item{polCs}{Catch: sd prior log normal parameter on relative catch}
\item{NK}{Number of nodes in the Gauss-Laguerre quadrature}
\item{gl_nodes}{Gauss-Laguerre quadrature nodes}
\item{gl_weights}{Gauss-Laguerre quadrature weights}
}
}
\source{
Caribbean Regional Fisheries Meeting Scientific Meeting 2007
}
\usage{
seatrout_ld
}
\description{
A data list created using the function \link{blicc_dat} using old length
frequency data collected from trawl and gillnet fisheries used in
the the CRFM 2007 Scientific Meeting.
}
\keyword{datasets}
