% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{sel_dsnormal}
\alias{sel_dsnormal}
\title{Calculate a double-sided normal selectivity curve for a length vector}
\usage{
sel_dsnormal(LMP, Smax, Ss1, Ss2, pstream__ = <pointer: 0x00007ffea6d9df40>)
}
\arguments{
\item{LMP}{A vector of lengths (usually mid-points for length bins)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A vector of selectivity values varying from 0.0 to 1.0
}\if{html}{\out{</div>}}
}
\description{
Place holder for the rstan::expose_stan_functions
}
