% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Stan.R
\name{set_LH_param}
\alias{set_LH_param}
\title{Sets data object from \code{blicc_dat()} to have a new selectivity functions}
\usage{
set_LH_param(blicc_ld, a = 1, b = 3, L50 = NA, L95 = NA, ma_L = NA, wt_L = NA)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model
(see function \code{blicc_dat()})}

\item{a}{The length-weight parameter: a*L^b. Not used in model fitting,
but for plots etc. Optional.}

\item{b}{The length-weight exponent (a*L^b, usually close
to 3.0) Optional.}

\item{L50}{The length at 50\% maturity, often referred to as "first"
maturity, and primarily applies to females. Must be less than Linf, usually
around 0.66 Linf, which is assumed if it is not provided. Optional.}

\item{L95}{The length at 95\% maturity. Must be greater than L50 and less
than Linf. A small increment is added L50 if it is not provided. Optional.}

\item{ma_L}{Mature biomass for each length bin. Optional.}

\item{wt_L}{Weight (biomass) for each length bin. Optional.}
}
\value{
The data object blicc_ld but with the new selectivities
}
\description{
Checks the gear and selectivity are valid, then sets up the data object
so that the parameters and parameter references for the selectivity
functions are valid. New references are replaced in the data object,
which is then returned.
}
\examples{
new_ld <- set_LH_param(eg_ld, a=1e-5, b=3.05, L50=20)

}
