% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Stan.R
\name{set_selectivity}
\alias{set_selectivity}
\title{Sets data object from \code{blicc_dat()} to have a new selectivity functions}
\usage{
set_selectivity(blicc_ld, Gear, sel_fun)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model
(see function \code{blicc_dat()})}

\item{Gear}{The gears for which the selectivity functions are being changed}

\item{sel_fun}{The selectivity functions that gears are being set to either as
integers or strings.}
}
\value{
The data object blicc_ld but with the new selectivities
}
\description{
Checks the gear and selectivity are valid, then sets up the data object
so that the parameters and parameter references for the selectivity
functions are valid. New references are replaced in the data object,
which is then returned.
}
\examples{
new_ld <- set_selectivity(eg_ld, Gear=1, sel_fun="logistic")

}
