% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_PR.R
\name{SMY_solve}
\alias{SMY_solve}
\title{Solve for a selectivity mode which produces maximum yield}
\usage{
SMY_solve(Fk, Linf, Galpha, Mk, Ss1, Ss2, blicc_ld, glq)
}
\arguments{
\item{Fk}{Fishing mortality (time in units of the growth rate K) to
be multiplied by selectivity to get fishing mortality in each length bin}

\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{blicc_ld}{A standard data list created by \code{blicc_dat}}

\item{glq}{a list of \code{nodes} and \code{weights} for the Gauss-Laguerre rule
(see \code{statmod::gauss.quad})}
}
\value{
maximum yield point for the selectivity mode
}
\description{
The function finds the selectivity mode (Smx) that maximises the yield per
recruit (YPR), keeping all other parameters at their current value.
}
\details{
Calculates yield per recruit.
This should generally work because there must be a maximum yield between
the extreme lengths (0 and Linf).
}
