% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_PR.R
\name{SSPR_solve}
\alias{SSPR_solve}
\title{Solve for a selectivity mode which produces the target SPR}
\usage{
SSPR_solve(Fk, Linf, Galpha, Mk, Ss1, Ss2, tarSPR, blicc_ld, glq)
}
\arguments{
\item{Fk}{Fishing mortality (time in units of the growth rate K) to
be multiplied by selectivity to get fishing mortality in each length bin}

\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Ss1}{Left side slope, parameterized as 1/sigma^2}

\item{Ss2}{Right side slope, parameterized as 1/sigma^2. Zero implies
flat-topped selectivity}

\item{tarSPR}{target SPR reference point: usually 0.2, 0.3 or 0.4}

\item{blicc_ld}{A standard data list created by \code{blicc_dat}}

\item{glq}{a list of \code{nodes} and \code{weights} for the Gauss-Laguerre rule
(see \code{statmod::gauss.quad})}
}
\value{
Selectivity mode (full selectivity) achieving the target SPR.
NA indicates this target cannot be achieved.
}
\description{
The function finds the selectivity mode (Smx) that gives the \code{tarSPR}
spawning potential ratio (SPR).
}
\details{
Calculates spawning biomass per recruit.
This should often work for sensible reference point target. However, note
that selectivity may not achieve any particular reference point if the
fishing mortality is too low. In these cases, NA is returned.
}
