% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{Survival_Est}
\alias{Survival_Est}
\title{Calculate the survival of a fish cohort to sequential length boundaries}
\usage{
Survival_Est(
  gl_node,
  gl_wt,
  Len,
  Zki,
  alpha,
  beta,
  pstream__ = <pointer: 0x00007ffea6d9df40>
)
}
\arguments{
\item{Len}{Vector of lower-bounds for length intervals}

\item{Zki}{Vector of total mortality (in time units of growth rate
parameter K) to be applied in each length interval.
This vector should be the same length as Len.}
}
\value{
A vector of proportion surviving to each length bin lower bound.
}
\description{
Place holder for the rstan::expose_stan_functions
}
