% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trgl_rp}
\alias{trgl_rp}
\title{A list of results including reference points from the \link{blicc_ref_pts} function}
\format{
A list of data frames of results and input information for a BLICC model fit:
\describe{
\item{dr_df}{A draws tibble of parameter estimates}
\item{lx_df}{A tibble of expected numbers of fish in each length bin for each gear}
\item{ld}{The data list from \link{blicc_dat} used to fit the model and calculate indicators and results}
\item{mcmc_summary}{Summary tibble of the Stan MCMC model fit}
\item{scenario}{A list of time period, gears, a vector describing the "direction" across gears and data list used to calculate reference points}
\item{rp_df}{A draws tibble of parameter estimates and reference points for the scenario}
}

dr_df is a tibble of parameter estimates from a BLICC model fit:
\describe{
\item{Linf}{Mean maximum length from the von Bertalanffy growth model, by growth group (list column)}
\item{Galpha}{Gamma distribution parameter governing growth variability}
\item{Mk}{Natural mortality (per unit K time), by growth group (list column)}
\item{Fk}{Fishing mortality, by gear and population (per unit K time) (list column)}
\item{Sm}{Selectivity model parameters (list column)}
\item{NB_phi}{Negative binomial parameter: Excess variance compared to the Poisson}
\item{lp__}{posterior log probability for the random draw}
\item{Gbeta}{Gamma distribution "rate" parameter, by growth group (=Galpha/Linf) (list column)}
\item{SPR}{Spawning potential ratio, by population (list column)}
\item{B_B0}{Harvestable biomass level relative to the unfished state, by population (list column)}
\item{YPR}{Yield per recruit, by gear (list column)}
\item{.chain}{MCMC chain identifier}
\item{.iteration}{MCMC iteration identifier}
\item{.draw}{MCMC draw identifier}
}

lx_df is a tibble of expected frequencies corresponding to the observed length frequency:
\describe{
\item{Qgroup}{Identifier for the observed data frequency}
\item{Lgroup}{Length lower bound as a factor}
\item{sel}{Selectivity estimate}
\item{N_L}{Population relative numbers at length estimate}
\item{efq}{Expected number of fish in the length bin}
}

rp_df is a tibble of reference points based on the scenario from a BLICC model fit:
\describe{
\item{Linf}{Mean maximum length from the von Bertalanffy growth model}
\item{Galpha}{Gamma distribution parameter governing growth variability}
\item{Mk}{Natural mortality (per unit K time)}
\item{Fk}{Fishing mortality by gear (per unit K time) (list column)}
\item{Sm}{Selectivity model parameters (list column)}
\item{NB_phi}{Negative binomial parameter: Excess variance compared to the Poisson}
\item{Gbeta}{Gamma distribution "rate" parameter (=Galpha/Linf)}
\item{.chain}{MCMC chain identifier}
\item{.iteration}{MCMC iteration identifier}
\item{.draw}{MCMC draw identifier}
\item{F20}{Fishing mortality estimated to achieve 20\% SPR (list column)}
\item{F40}{Fishing mortality estimated to achieve 40\% SPR (list column)}
\item{S40}{Selectivity location estimated to achieve 40\% SPR (list column)}
\item{SMY}{Selectivity location estimated to achieve maximum yield per recruit (list column)}
}
}
\source{
Data are simulated.
}
\usage{
trgl_rp
}
\description{
An example data table of parameter draws and reference points from the BLICC model
trgl_rp <- blicc_ref_pts(trgl_slim, trgl_ld)
}
\keyword{datasets}
