% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{weight_at_length}
\alias{weight_at_length}
\title{Weight at length}
\usage{
weight_at_length(blicc_ld)
}
\arguments{
\item{blicc_ld}{A standard data list created by \code{blicc_dat}}
}
\description{
The weight is calculated using the standard length-weight model W=a L^b.
In terms of reference points only the exponent parameter b is required,
and the parameter a is only used for scaling.
}
\examples{
ld <- blicc_dat(LLB = 10:35, a=1.0e-4, b=2.95, L50=24,
                fq=c(0,1,2,5,26,70,72,66,36,34,25,24,
                     20,10,12,5,3,5,6,4,2,0,2,0,1,0), Linf=c(32, 3))
wt <- weight_at_length(ld)
plot(y=wt, x=ld$LMP, type="l")

}
