% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blicc_dat}
\alias{blicc_dat}
\title{Generates a data list, including first-guess suitable parameter values, for the BLICC model}
\usage{
blicc_dat(LLB, fq, Linf, lMk = NA, b = 3, L50 = 0.66, Ls = 0.05)
}
\arguments{
\item{LLB}{A vector of the lower length boundaries for the length frequency}

\item{fq}{A vector of the same length as LLB containing the frequency data. Zeroes must be included.}

\item{Linf}{A vector of two values being the mean and sd of the prior mean maximum length for the species}

\item{lMk}{Log natural mortality divided by the growth rate K (usually around 1.5)}

\item{b}{A number with the length-weight exponent (usually close to 3.0)}

\item{L50}{A number for the length at 50\% maturity as a proportion of Linf (primarily applies to females). Must be less than 1.0.}

\item{Ls}{A number for the logistic steepness parameter for the maturity ogive}

\item{bw}{A number providing the bin width (usually 1)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A list of vectors and numbers structured suitable for use in the Stan model BLICC.stan
}\if{html}{\out{</div>}}
}
\description{
Vectors for the lower bound for each length bin and the number of fish in each length bin and combined with
prior parameters into a data list as expected by the BLICC model. Constant bin widths are
not necessary for the model and not assumed. Zero frequencies must be explicitly in the fq vector.
}
