% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blicc_dat}
\alias{blicc_dat}
\title{Generates a data list and initial parameter values for the BLICC model}
\usage{
blicc_dat(
  model_name = "fishblicc data model",
  LLB,
  fq,
  Linf,
  sel_fun,
  gear_sel = NULL,
  Catch = 1,
  gear_names = NA,
  Mk = NA,
  ref_length = -1,
  wt_L = NA,
  ma_L = NA,
  a = 0.001,
  b = 3,
  L50 = NA,
  L95 = NA,
  NK = NA
)
}
\arguments{
\item{model_name}{A name for the model (species or fishery). Optional.}

\item{LLB}{A vector of lower length boundaries for each length frequency
bin. Required.}

\item{fq}{A list of vectors, each the same length as LLB, containing the
frequency data for each gear. Zeroes must be included. If only one gear,
can be a vector. Required.}

\item{Linf}{A vector of two values: mean and sd of the prior maximum mean
length for the stock. Required.}

\item{sel_fun}{A vector of selectivity function names (character) or index
(1-4), with 1="logistic", 2="normal", 3="ss_normal", 4="ds_normal". Length
must be the same as fq if \code{gear_sel} is not provided. Required.}

\item{gear_sel}{A list of integer vectors indexing which \code{sel_fun} are used
for which gear. It can include multiple selectivity functions for each
gear.  Optional.}

\item{Catch}{A vector of relative catches, one for each gear. Required if the
number of gears is more than one.}

\item{gear_names}{A vector of names for the gears for reference. Optional.}

\item{Mk}{Natural mortality divided by the growth rate K (usually around
1.5). Optional.}

\item{ref_length}{The reference length for the inverse length function if it
is used. (default is NA i.e. constant natural mortality)}

\item{wt_L}{Weight (biomass) for each length bin. Optional.}

\item{ma_L}{Mature biomass for each length bin. Optional.}

\item{a}{The length-weight parameter: a*L^b. Not used in model fitting, but
for plots etc. Optional.}

\item{b}{The length-weight exponent (a*L^b, usually close to 3.0) Optional.}

\item{L50}{The length at 50\% maturity, often referred to as "first" maturity,
and primarily applies to females. Must be less than Linf, usually around
0.66 Linf, which is assumed if it is not provided. Optional.}

\item{L95}{The length at 95\% maturity. Must be greater than L50 and less
than Linf. A small increment is added to L50 if it is not provided.
Optional.}

\item{NK}{Number of nodes for the Gauss Laguerre quadrature rule. 110 is a
safe value, but extends the run time of all calculations.  If not provided
it is estimated. Optional.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list structured in a manner suitable for use in the Stan model
}\if{html}{\out{</div>}}

\code{BLICC.stan}.
}
\description{
Vectors for the lower bound for each length bin and the number of fish in
each length bin and combined with prior parameters into a data list as
expected by the BLICC model.
}
\details{
This is a convenience function. The fitting and other functions
accept a linked list of values providing necessary information for the
BLICC model, including the length bin boundaries and length frequencies.
Some basic error checking is carried out and error message is given if the
inputs are incorrect.

The data list produced as output contains data and parameters suitable for
the model fit and producing output. These values can be changed in the
list, but must have the same format. The majority of information is the
length bin, frequency data and prior parameters (normal for Linf and
log-normal for all other parameters). However, the list also contains NK,
the number of knots (nodes) used in the quadrature integration. The default
is 110, which is safe but extends running times. Lower values risk
inaccurate integration but result in faster fitting.

The length frequencies should include zeros for bins which contained no
fish. The length frequency should be bounded by a single zero bin as the
first and last bin in the frequency.

To allow maximum flexibility, prior hyper-parameters for the selectivity
functions (as log-values for means in the log-normal distribution) are
defined in a single vector. These are linked to each function using an
integer matrix with a row for each selectivity function (gear)
}
\examples{
ld <- blicc_dat(LLB = 25:35,
                fq = list(c(0,1,2,26,72,66,36,24,12,4,0)),
                Linf = c(35, 2),
                sel_fun = 4,
                gear_names = "Gill net")

}
