% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_stan.R
\name{blicc_dat}
\alias{blicc_dat}
\title{Generates a data list and initial parameter values for the BLICC model}
\usage{
blicc_dat(
  LLB,
  fq,
  Linf,
  Mk = NA,
  a = 1,
  b = 3,
  L50 = NA,
  L95 = NA,
  Flat = 1L,
  NK = 110L
)
}
\arguments{
\item{LLB}{A vector of lower length boundaries for each length frequency
bin. Required.}

\item{fq}{A vector, the same length as LLB, containing the frequency data.
Zeroes must be included. Required.}

\item{Linf}{A vector of two values: mean and sd of the prior maximum mean
length for the stock. Required.}

\item{Mk}{Natural mortality divided by the growth rate K
(usually around 1.5). Optional.}

\item{a}{The length-weight parameter: a*L^b. Not used in model fitting,
but for plots etc. Optional.}

\item{b}{A number with the length-weight exponent (a*L^b, usually close
to 3.0) Optional.}

\item{L50}{The length at 50\% maturity, often referred to as "first"
maturity, and primarily applies to females. Must be less than Linf, usually
around 0.66 Linf, which is assumed if it is not provided. Optional.}

\item{L95}{The length at 95\% maturity. Must be greater than L50 and less
than Linf. A small increment is added L50 if it is not provided. Optional.}

\item{Flat}{An integer indicating whether selectivity is flat-topped
(==0) or not (!=0).  Optional.}

\item{NK}{Number of nodes for the Gauss Laguerre quadrature rule. 110 is a
safe value, but extends the run time of all calculations.  Optional.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list of vectors and numbers structured suitable for use in the
}\if{html}{\out{</div>}}

Stan model \code{BLICC.stan}
}
\description{
Vectors for the lower bound for each length bin and the number of fish in
each length bin and combined with prior parameters into a data list as
expected by the BLICC model. Constant bin widths are not necessary for
the model and not assumed. Zero frequencies must be explicitly in the
fq vector.
}
\details{
This is a convenience function. The fitting and other functions accept a
linked list of values providing necessary information for the BLICC model,
including the length bin boundaries and length frequencies. Some basic error
checking is carried out and error message is given if the inputs are
incorrect.

The data list produced as output contains data and parameters suitable for
the model fit and producing output. These values can be changed in the list,
but must have the same format. The majority of information is the length
bin, frequency data and prior parameters (normal for Linf and log-normal
for all other parameters). However, the list also contains NK, the number
of knots (nodes) used in the quadrature integration. The default is 110,
which is safe but extends running times. Lower values risk inaccurate
integration but result in faster fitting.

It is important that the length frequencies include zeros for bins which
contained no fish. The length frequency should be bounded by a single zero
bin as the first and last bin in the frequency.
}
\examples{
ld <- blicc_dat(LLB = 25:35, fq=c(0,1,2,26,72,66,36,24,12,4,0), Linf=c(35, 2))

}
