% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{blicc_get_expected}
\alias{blicc_get_expected}
\title{Expected length interval selectivity, mortality, survival and frequency}
\usage{
blicc_get_expected(Linf, Galpha, Mk, Fk, Sm, blicc_ld)
}
\arguments{
\item{Linf}{Maximum mean length fish in the population can grow to}

\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Mk}{Natural mortality (time in units of the growth rate K)}

\item{Fk}{Vector of fishing mortality (time in units of the growth
rate K) to be multiplied by selectivity to get fishing mortality in each
length bin}

\item{Sm}{Vector of parameters for all the selectivity functions
flat-topped selectivity}

\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}
}
\value{
A tibble of expected values for each length bin for each gear
}
\description{
For a set of parameter values, returns the expected selectivity, mortality,
survival, and catch in each length bin based on the BLICC model as a tibble.
}
\details{
The model and provided parameter values are used to calculate the expected
values for selectivity, survival, relative population numbers,
total mortality and expected length frequency for each length bin.
Input parameters are point estimates (reals). The standard data list
provides the model dimensions and is produced by the function
\code{\link[=blicc_dat]{blicc_dat()}}.
}
\examples{
blicc_get_expected(Linf = 32, Galpha = 100, Mk = 1.5, Fk = 1.5,
                   Sm = c(24, 0.1, 0.001), blicc_ld = eg_ld)

}
