% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{blicc_ref_pts}
\alias{blicc_ref_pts}
\title{Generate posterior draws and standard reference points}
\usage{
blicc_ref_pts(slimf, blicc_ld, vdir = NULL, a = NA, b = NA, L50 = NA, L95 = NA)
}
\arguments{
\item{slimf}{An object from the \link{blicc_fit} or \link{blicc_mpd} functions.}

\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{vdir}{A search direction vector with maximum value 1 and minimum 0
applied to changes across gears. Optional.}

\item{a}{The length-weight parameter: a*L^b. Not used in model fitting, but
used to calculate \code{wt_L} if that is not provided. for plots etc. Optional.}

\item{b}{The length-weight exponent (a*L^b, usually close to 3.0). Not used
in model fitting, but used to calculate \code{wt_L} if that is not provided.
Optional.}

\item{L50}{The length at 50\% maturity, often referred to as "first" maturity,
and primarily applies to females. Must be less than \code{Linf}, usually around
0.66 Linf, which is assumed if it is not provided. Optional.}

\item{L95}{The length at 95\% maturity. Must be greater than L50 and less
than Linf. A small increment is added to L50 if it is not provided.
Optional.}
}
\value{
A list of 1) the search direction vector 2) posterior draws or mpd
point estimates of important parameters and per-recruit reference points 3)
observed-expected length frequencies, 4) \code{blicc_ld} data object used.
}
\description{
The function returns a list consisting of the search direction vector, a
tibble of draws from the posterior MCMC stanfit (or the mpd estimate)
including SPR 20\%, 30\% and 40\% reference points, expected number in each
length bin and the data object used to generate these.  The tibble of draws
is suitable for use in the posterior and bayesplot packages and can be used
in various functions to show results. Note that the function can take a
little time to run and may be moderately large, depending on the number of
draws. Some reference points may not exist (indicated by \code{NA}).
}
\details{
Reference points are found where possible for the parameter draw
from the MCMC. The reference points are found along a line representing the
relative impact across gears. The default is to apply the same adjustments
proportionately to every gear that has a fishing mortality, but alternative
scenarios can be defined. For example, for two gears the default is
\code{vdir=c(1, 1)}. if there are two gears, and only the first will be
adjusted, then the direction would be \code{vdir=c(1, 0)}. It is possible to
provide a vector such as \code{vdir=c(1, 0.5)} to represent only a 50\% reduction
in the second gear compared to the first, for example. This will only be
applied to the fishing mortality. The selectivity changes will continue to
use the vector as a dummy variable, but the quantity will be ignored and
all \code{vdir} with non-zero values for gears will be treated as having equal
displacement; so for example a selectivity evaluation for
\code{vdir = c(1.0, 0.5)} is equivalent to \code{vdir = c(1, 1)}.

It is likely that proper simulation projections would be required for more
complex cases where gears are not being simply excluded from management
action. Unless there is a specific need, the best option is to leave it as
the default \code{NULL} so the reference points are indicative and consistent.

Where reference points do not exist an \code{NA} is returned. It is quite
possible reference points do not exist in many cases due to the effects of
selectivity. Strictly speaking, reference points may not exist within
specific constraints on F and selectivity. Within more general bounds on F
and selectivity, all reference "points" exist as lines or points. In these
functions, reference points in relation to a control (\code{Fk} or \code{Sm}) are
calculated based on the other being fixed at the current estimated value.
To see how the fishery might respond when they are adjusted together, the
yield or SPR surface must be plotted.
}
\examples{
\dontrun{
res_rp <- blicc_ref_pts(eg_slim, eg_ld)
summary(res_rp$rp_df)
}

}
