% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{blicc_ref_pts}
\alias{blicc_ref_pts}
\title{Posterior draws including standard reference points for F and Smx}
\usage{
blicc_ref_pts(slimf, blicc_ld)
}
\arguments{
\item{slimf}{A stanfit object from the \code{blicc_fit} function.}

\item{blicc_ld}{A data list from the \code{blicc_dat} function.}
}
\value{
A tibble containing posterior draws of important parameters and
per-recruit reference points
}
\description{
The functions returns a tibble of draws from the posterior MCMC stanfit
object together with a nested tibble for each draw containing the results
(expected values) in terms of length. The tibble is suitable for use in
the \code{posterior} and \code{bayesplot} packages and can be used in various
functions to show results. Note that the function can take a little time
to run and may be moderately large, depending on the number of
draws, and reference points may not exist (indicated by \code{NA}).
}
\details{
Reference points are found where possible for parameter draw from the MCMC.
Where reference points do not exist an \code{NA} is returned. It is quite
possible reference points do not exist in many cases due to the effects of
selectivity. Strictly speaking, reference points may not exist within
specific constraints on F and selectivity. Within more general bounds on F
and selectivity, all reference "points" exist as lines or points. In these
functions, reference points in relation to a control (Fk or Smx) are
calculated based on the other being fixed at the current estimated value.
To see how the fishery might respond when they are adjusted together,
the yield or SPR surface must be plotted.
}
\examples{
\dontrun{
rp_df <- blicc_ref_pts(eg_slim, eg_ld)
summary(rp_df)
}

}
