% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blicc_selfun}
\alias{blicc_selfun}
\title{Sets data object from \link{blicc_dat} to have new selectivity functions}
\usage{
blicc_selfun(blicc_ld, sel_fun, sel_indx = NULL, model_name = NULL)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{sel_fun}{A vector of selectivity function names (character) or index
(1-4), with 1="logistic", 2="normal", 3="ss_normal", 4="ds_normal". Length
must be the same as fq if \code{gear_sel} is not provided. Required.}

\item{sel_indx}{An integer vector index of the selectivity functions to be
changed. If not provided, all functions will be replaced.}

\item{model_name}{A name for the model (species or fishery). Optional.}
}
\value{
The data object \code{blicc_ld} with the new selectivities
}
\description{
Checks selectivity functions are valid, then updates the data object with the
new functions so that the parameters and parameter references for the
selectivity functions are valid. New references are replaced in the data
object, which is then returned.
}
\details{
Particular selectivity functions can be changed using the \code{sel_indx} parameter,
which indexes which functions will be changed. In this case, only those
functions will need to have the prior parameters updated.
}
\examples{
new_ld <- blicc_selfun(eg_ld, sel_fun="logistic", model_name = "Logistic Selectivity")

}
