% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Stan.R
\name{blip_Fk}
\alias{blip_Fk}
\title{Sets data object from \code{\link[=blicc_dat]{blicc_dat()}} to have a new Fk prior}
\usage{
blip_Fk(blicc_ld, lFk = NA, lFks = NA, model_name = NA)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{lFk}{A vector of double containing the lognormal mean Fk for each gear}

\item{lFks}{A double containing the lognormal sd Fk for all gears}

\item{model_name}{A string for a replacement model name in the data
object. Optional.}
}
\value{
The data object \code{blicc_ld} with the prior for \code{Fk} changed.
}
\description{
The Fk lognormal prior is updated with new mu and a single sigma parameter.
If no value (\code{NA}) is provided for lFk, the defaults are assigned. These are
the current natural mortality mu for fishing mortality mu and 2.0 for the
sigma parameter.
}
\examples{
new_ld <- blip_Fk(eg_ld, lFk=log(1.9), lFks=1.5)

}
