% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_Galpha}
\alias{blip_Galpha}
\title{Sets data object from \code{\link[=blicc_dat]{blicc_dat()}} to have a new Galpha prior}
\usage{
blip_Galpha(blicc_ld, lGalpha)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{lGalpha}{A numeric vector of double containing the mean and sd for the
prior log-normal.}
}
\value{
The data object blicc_ld but with the prior for Galpha changed.
}
\description{
The Galpha prior is updated with new values.
CV: could be 5\% (log(1/0.05^2)) to 30\% (log(1/0.3^2)).
A 30\% CV makes length very uninformative on age however.
Default CV=10\%
}
