% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_LH}
\alias{blip_LH}
\title{Set life-history parameters and vectors in the data object}
\usage{
blip_LH(
  blicc_ld,
  a = NULL,
  b = NULL,
  L50 = NULL,
  L95 = NULL,
  ma_L = NULL,
  wt_L = NULL,
  model_name = NULL,
  set_defaults = FALSE
)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{a}{The length-weight parameter: a*L^b. Not used in model fitting, but
used to calculate \code{wt_L} if that is not provided. for plots etc. Optional.}

\item{b}{The length-weight exponent (a*L^b, usually close to 3.0). Not used
in model fitting, but used to calculate \code{wt_L} if that is not provided.
Optional.}

\item{L50}{The length at 50\% maturity, often referred to as "first" maturity,
and primarily applies to females. Must be less than \code{Linf}, usually around
0.66 Linf, which is assumed if it is not provided. Optional.}

\item{L95}{The length at 95\% maturity. Must be greater than L50 and less
than Linf. A small increment is added to L50 if it is not provided.
Optional.}

\item{ma_L}{Mature biomass for each length bin. Optional.}

\item{wt_L}{Weight (biomass) for each length bin. Optional.}

\item{model_name}{A string for a replacement model name in the data object.
Optional.}

\item{set_defaults}{Logical indicating whether to set defaults if parameters
are \code{NA}. Leaves them alone if \code{FALSE}.}
}
\value{
The data object blicc_ld with the new life-history parameters
}
\description{
Sets the maturity-at-length and weight-at-length vectors in the data object
(from \link{blicc_dat}) to new values. The values are either provided as vectors
or model parameters to calculate the vectors. Calculations are done for the
length bin mid-points. The models for the calculation are the standard
length-weight (\code{W=aL^b}) and logistic model for the maturity-at-length: (\code{Lm = 1/(1+exp(Ls*(L-L50)))}).
}
\examples{
new_ld <- blip_LH(eg_ld, a=1.2e-5, b=2.95, model_name="Alternative LW")

}
