% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_main_sel}
\alias{blip_main_sel}
\title{Set gear selectivity component to main}
\usage{
blip_main_sel(blicc_ld, gear, sel_N)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{gear}{Single integer or exact name indexing one gear}

\item{sel_N}{Index for the new main selectivity component with weight=1.}
}
\value{
The data object \code{blicc_ld} but with the new main selectivity
}
\description{
A gear's selectivity component is set so that its weight equals 1.0. Other
mixture weights are adjusted accordingly.
}
