% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Stan.R
\name{blip_Mk}
\alias{blip_Mk}
\title{Sets data object from \code{\link[=blicc_dat]{blicc_dat()}} to have a new Mk prior}
\usage{
blip_Mk(
  blicc_ld,
  lMk = as.numeric(c(NA, NA)),
  ref_length = NA,
  model_name = NA
)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{lMk}{The (log) mean and sigma for the lognormal natural mortality prior}

\item{ref_length}{The reference length for the inverse length function if it
is used. (default is NA i.e. fixed natural mortality)}

\item{model_name}{A string for a replacement model name in the data
object. Optional.}
}
\value{
The data object blicc_ld but with the prior and function for Mk
changed.
}
\description{
Checks the new values are valid, then provide the mean and sd for
the log-normal prior to be used. Also, if the ref_length parameter
is defined, apply the inverse length function for natural mortality.
These are replaced in the data object, which is then returned.
Note that the Mk must be provided as the log value.
}
\examples{
new_ld <- blip_Mk(eg_ld, lMk=c(log(1.9), 0.2), ref_length=25)

}
