% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_blip.R
\name{blip_NBphi}
\alias{blip_NBphi}
\title{Sets data object from \link{blicc_dat} to have a new \code{NB_phi} prior}
\usage{
blip_NBphi(blicc_ld, lNBphi)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{lNBphi}{A numeric vector of double containing the mu and sigma for the
prior log-normal.}
}
\value{
The data object blicc_ld but with the prior for \code{NBphi} changed.
}
\description{
The \code{NB_phi} lognormal prior is updated with new values mu and sigma.
NB_phi controls the negative binomial overdispersion compared to the
Poisson distribution, where the variance is \code{mu + mu^2 / NB_phi}.
}
