% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blicc_stan.R
\name{blip_NBphi}
\alias{blip_NBphi}
\title{Sets data object from \code{\link[=blicc_dat]{blicc_dat()}} to have a new NB_phi prior}
\usage{
blip_NBphi(blicc_ld, lNBphi)
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{lNBphi}{A numeric vector of double containing the mean and sd for the
prior log-normal.}
}
\value{
The data object blicc_ld but with the prior for Galpha changed.
}
\description{
The NB_phi lognormal prior is updated with new values mu and sigma.
NB_phi controls the negative binomial overdispersion compared to the
Poisson distribution, where the variance is mu + mu^2 / NB_phi.
}
