% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_ld}
\alias{eg_ld}
\title{Example small sample of length frequency data}
\format{
A list of 26 data vectors used in the BLICC model:
\describe{
\item{NK}{Number of nodes in the Gauss-Laguerre quadrature}
\item{oBN}{Number of bins in the length frequency}
\item{Len}{Vector of length bin lower boundaries}
\item{LMP}{Length bin mid-points, used for selectivity and plotting}
\item{fq}{Vector of length frequencies}
\item{a}{Length-weight parameter in a L^b. Optional}
\item{b}{Length-weight parameter in a L^b.}
\item{Ls}{Logistic steepness parameter for the maturity ogive}
\item{Lm}{Logistic parameter for 50\% maturity}
\item{Flat_top}{0 == Flat-topped selectivty; Otherwise selectivity is domed}
\item{poLinfm}{Linf: Mean prior normal parameter}
\item{poLinfs}{Linf: sd prior normal parameter}
\item{polGam}{Galpha: Mean prior normal log parameter}
\item{polGas}{Galpha: sd prior normal log parameter}
\item{polMkm}{Mk: Mean prior normal log parameter}
\item{polMks}{Mk: sd prior normal log parameter}
\item{polFkm}{Fk: Mean prior normal log parameter}
\item{polFks}{Fk: sd prior normal log parameter}
\item{polSmxm}{Smx: Mean prior normal log parameter}
\item{polSmxs}{Smx: sd prior normal log parameter}
\item{polSs1m}{Ss1: Mean prior normal log parameter}
\item{polSs1s}{Ss1: sd prior normal log parameter}
\item{polSs2m}{Ss2: Mean prior normal log parameter}
\item{polSs2s}{Ss2: sd prior normal log parameter}
\item{polNB_phim}{NB_phi: Mean prior log normal parameter}
\item{polNB_phis}{NB_phi: sd prior log normal parameter}
}
}
\source{
Data are simulated.
}
\usage{
eg_ld
}
\description{
A data list created using simulated length frequency the function \code{blicc_dat}:
eg_ld <- blicc_dat(LLB = 10:35, a=1.0e-4, b=2.95, L50=24,
fq=c(0,1,2,5,26,70,72,66,36,34,25,24,20,10,12,5,3,5,6,4,2,0,2,0,1,0),
Linf=c(32, 3))
}
\keyword{datasets}
