% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishblicc-package.R
\docType{package}
\name{fishblicc-package}
\alias{fishblicc-package}
\alias{fishblicc}
\title{The 'fishblicc' package.}
\description{
Procedures to fit catch curves to length frequency data based using length intervals.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.13. https://mc-stan.org
}
