% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_efq_FRP}
\alias{plot_efq_FRP}
\title{Plot expected length frequencies for a range of fishing mortality
reference points}
\usage{
plot_efq_FRP(blicc_rp, blicc_lx, blicc_ld)
}
\arguments{
\item{blicc_rp}{A posterior draws and reference points tibble
from \code{blicc_ref_pts} function.}

\item{blicc_lx}{A tibble from \code{blicc_expect_len} function.}

\item{blicc_ld}{A data list from the \code{blicc_dat} function.}
}
\value{
ggplot geom object for plotting expected length frequencies by SPR
}
\description{
The facet plot covers the current estimated fishing mortality, together with
fishing mortalities required to obtain SPR 20\%, SPR 30\% and SPR 40\%. The
graphs show the expected values and 80\% credible intervals compared to
current observations, and can be used to assess whether length frequencies
should be able to detect changes in fishing mortality to these different
levels.
}
\examples{
eg_lx <- blicc_expect_len(eg_rp, eg_ld)
plot_efq_FRP(eg_rp, eg_lx, eg_ld)

}
