% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_efq_FRP}
\alias{plot_efq_FRP}
\title{Plot expected length frequencies for a range of fishing mortality reference
points}
\usage{
plot_efq_FRP(blicc_rp, gear = NULL)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{gear}{Specifies the gear to plot as an integer index or full name}
}
\value{
ggplot geom object for plotting expected length frequencies by SPR
level
}
\description{
The facet plot covers the current estimated fishing mortality, together with
fishing mortalities required to obtain SPR 20\%, SPR 30\% and SPR 40\%. The
graphs show the expected values and 80\% credible intervals compared to
current observations, and can be used to assess whether length frequencies
should be able to detect changes in fishing mortality to these different
levels. Note that if reference points do not exist, they are not plotted, so
some graphs may be blank except for the data. The graphs show the number of
MCMC draws for which the reference points exist in each case.
}
\examples{
plot_efq_FRP(eg_rp)

}
