% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_efq_SRP}
\alias{plot_efq_SRP}
\title{Plot expected length frequencies for a range of selectivity reference points}
\usage{
plot_efq_SRP(blicc_rp, blicc_lx, blicc_ld)
}
\arguments{
\item{blicc_rp}{A posterior draws and reference points tibble
from \code{blicc_ref_pts} function.}

\item{blicc_lx}{A tibble from \code{blicc_expect_len} function.}

\item{blicc_ld}{A data list from the \code{blicc_dat} function.}
}
\value{
ggplot geom object plotting expected length frequencies by
selectivity
}
\description{
The facet plot covers the current estimated selectivity mode (Smx), together
with selectivity changes required to obtain SPR 20\%, SPR 40\% and the maximum
yield. The graphs show the expected values and 80\% credible intervals
compared to current observations, and can be used to assess whether length
frequencies should be able to detect changes in selectivity to these
different levels.
}
\examples{
eg_lx <- blicc_expect_len(eg_rp, eg_ld)
plot_efq_SRP(eg_rp, eg_lx, eg_ld)

}
