% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_expected_frequency}
\alias{plot_expected_frequency}
\title{Plot the observed and expected frequency data}
\usage{
plot_expected_frequency(blicc_rp, gear = NA)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{gear}{Specifies the gear to plot as an integer index or full name}
}
\value{
ggplot geom object plotting observed and expected frequency
}
\description{
The graph shows the observed frequency and the expected frequency for the
specified gears. The plots include the 80\% credible intervals for the
estimate (if an MCMC is provided) and 80\% credible intervals for the
observations. The latter takes into account the negative binomial observation
error, so on average 8 out of ten observations would be expected within these
bounds.
}
\examples{
plot_expected_frequency(eg_rp)

}
