% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_prior}
\alias{plot_prior}
\title{Plot the observed and expected frequency data based on the prior parameters}
\usage{
plot_prior(blicc_ld, gear = "All", time_period = "All")
}
\arguments{
\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}

\item{gear}{Specifies the gear(s) to plot as an integer index or full name}

\item{time_period}{Specifies the period(s) to plot as an integer index or
full name}
}
\value{
ggplot geom object plotting observed and the prior's expected
frequency, separated by gear.
}
\description{
The graph shows the observed frequency and the prior's predicted frequency
for all (or specified) gears. This allows a check that current prior
parameter settings are consistent with observations. The prior means are also
used as the start point for the \link{blicc_mpd} fits.
}
\examples{
plot_prior(eg_ld)

}
