% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot standardised residuals against length}
\usage{
plot_residuals(blicc_rp, gear = "All")
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{gear}{Specifies the gear(s) to plot as an integer index or full name}
}
\value{
ggplot geom object plotting standardised residuals
}
\description{
The graph shows the standardised residuals plotted against length for all
gears combined, separated by colour. Due to the variance weighting, outliers
can most be expected at the tails of the frequency. For the MCMC output, the
residuals are plotted as 99\% quartiles using error bars, together with the
outliers as points outside this range.
}
\details{
If you want to alter the graph, the group names are \code{Sgroup} for the
selectivities and \code{Lgroup} for the lengths.
}
\examples{
plot_residuals(eg_rp)

}
