% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_selectivity}
\alias{plot_selectivity}
\title{Plot selectivity and 50\% and 80\% credible intervals by length}
\usage{
plot_selectivity(blicc_lx, blicc_ld)
}
\arguments{
\item{blicc_lx}{A tibble from \code{blicc_expect_len} function.}

\item{blicc_ld}{A data list from the \code{blicc_dat} function.}
}
\value{
ggplot geom object for plotting
}
\description{
Plot selectivity and 50\% and 80\% credible intervals by length
}
\examples{
eg_lx <- blicc_expect_len(eg_rp, eg_ld)
plot_selectivity(eg_lx, eg_ld)

}
