% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_SPR_density}
\alias{plot_SPR_density}
\title{Plot of the estimated spawning potential ratio probability density}
\usage{
plot_SPR_density(blicc_rp)
}
\arguments{
\item{blicc_rp}{A posterior draws and reference points tibble
from \code{blicc_ref_pts} function.}
}
\value{
ggplot geom object for plotting
}
\description{
Plot of the estimated spawning potential ratio probability density
}
\examples{
plot_SPR_density(eg_rp)

}
