% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Plots.R
\name{plot_YPR_contour}
\alias{plot_YPR_contour}
\title{Yield-per-recruit surface contour plot}
\usage{
plot_YPR_contour(blicc_rp, Gear = NA)
}
\arguments{
\item{blicc_rp}{A list of posterior draws, reference points with associated
direction, the data object and expected lengths from \link{blicc_ref_pts}
function.}

\item{Gear}{Specifies the gear to plot as an integer index or full name}
}
\value{
ggplot geom object for plotting
}
\description{
The plot shows the YPR surface plotted against the fishing mortality (Fk) and
the 50\% or maximum selectivity, dependent on the selectivity function. The
current state of the stock is marked as a point and the F0.1 reference point
is marked as a line. All values are calculated as means and uncertainty is
not represented to maintain simplicity. The reference line is plotted if it
exists within a reasonable range of the current estimated status. For
multiple gears, the surface is calculated for a single reference gear.
}
\examples{
plot_YPR_contour(eg_rp)

}
