% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_stan.R
\name{posterior_density_ratio}
\alias{posterior_density_ratio}
\title{Calculates the mean log-probability density ratio between two sets of draws}
\usage{
posterior_density_ratio(stf1, stf0)
}
\arguments{
\item{stf1, stf0}{Stan model output for the numerator and denominator}
}
\value{
A single number estimating the relative probability in support for
model stf1 compared to stf2
}
\description{
Simple Bayes factor estimate based on the mean posterior density.
This only works where data and likelihood model remain the same in stf1
and stf0. The only change is a single fixed parameter in stf0 is estimated
in stf1. This allows direct comparison of the estimated total posterior
probability between the two fits. This can be used to assess the fit
between a domed and flat-topped selectivity function.
}
\examples{
\dontrun{
ld <- blicc_dat(LLB = 25:35, fq=c(0,1,2,26,72,66,36,24,12,4,0),
                Linf=c(35, 2))
slim_domed <- blicc_fit(ld)
ld$Flat <- 0L  # Make selectivity flat-topped (logistic or knife-edged)
slim_flat <- blicc_fit(ld)
posterior_density_ratio(slim_domed, slim_flat)
}
}
