% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_ObsExp.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Get a posterior predicted length frequency}
\usage{
posterior_predict(blicc_rp, gear = NULL, draws = 0)
}
\arguments{
\item{blicc_rp}{List of fishblicc result tables from \link{blicc_ref_pts}}

\item{gear}{Identifies the single gear (selectivity) providing predictions}

\item{draws}{The number of random draws up to the number of draws from the
MCMC (the default).}
}
\value{
A matrix with rows equal to gears*draws and columns to length
}
\description{
For a set of MCMC parameter draws, returns a simulated catch in each
length bin based on the BLICC model. This matrix of simulated data that
might be expected if the model is correct, can be used in the posterior
and \code{bayesplot} packages for evaluation.
}
\examples{
yrep <- posterior_predict(blicc_rp = eg_rp, gear=1, draws=100)

}
