% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{Rpop_F}
\alias{Rpop_F}
\title{Calculate the survival of a fish cohort to sequential length boundaries}
\usage{
Rpop_F(Galpha, Gbeta, Mk, Fk, FSel, blicc_ld)
}
\arguments{
\item{Galpha}{Alpha parameter for the Gamma probability density function
that governs growth variability.}

\item{Gbeta}{Rate parameter for the Gamma distribution growth variability
(=Galpha/Linf)}

\item{Mk}{Natural mortality divided by the growth rate K}

\item{Fk}{Fishing mortality divided by the growth rate K for each gear
making a contribution}

\item{FSel}{A list of all the selectivities for each length bin}

\item{blicc_ld}{A standard data list suitable for the model
(see function \code{blicc_dat()})}
}
\value{
A list of the population size in each length bin and a list of
fishing mortalities at length for each gear.
}
\description{
The model integrates over growth rate variability and sums mortality
piece-wise over intervening length intervals to derive the
survival to each length bin boundary taking into account variation in growth
and varying mortality-at-length. This is analogous to an age based catch
curve model, but for length. Survival can be used to estimate population
numbers-at-length and catch length frequency, among other things.
The model uses the Gauss-Laguerre quadrature rule for integration,
so nodes and weights for this must be provided.
}
\examples{
Sel <- Rselectivities(exp(eg_ld$polSm), eg_ld)
S <- Rpop_F(100, 100/50, Mk=1.5, Fk=exp(eg_ld$polFkm),
            FSel=Sel, blicc_ld=eg_ld)
plot(y=S$NL, x=eg_ld$LMP, type="l")

}
