% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLICC_Model.R
\name{Rselectivities}
\alias{Rselectivities}
\title{Calculate the selectivities for each length bin for each gear}
\usage{
Rselectivities(Sm, blicc_ld)
}
\arguments{
\item{Sm}{Vector of parameters for all the selectivity functions
including mixture weights at the end of the vector}

\item{blicc_ld}{A standard data list suitable for the model (see function
\link{blicc_dat})}
}
\value{
Selectivities as a vector for each length bin in a list for each
gear.
}
\description{
The function calculates the selectivities for each gear based on the relevant
selectivity functions and their parameters.
}
\examples{
S <- Rselectivities(exp(eg_ld$polSm), eg_ld)
plot(y=S[[1]], x=eg_ld$LMP, type="l")
}
